requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.2
pytz~=2024.2
mt5-grpc-proto==0.1.0